+++ Banter/Infrastructure/Web3Client/Adapter/Web3AsyncAdapter.swift
@@ -11,77 +11,6 @@ import Web3
 import Web3ContractABI
 import Web3PromiseKit
 
-protocol Web3Async {
-  func getTransactionCount(
-    for address: EthereumAddress, block: EthereumQuantityTag
-  ) async throws -> EthereumQuantity
-
-  func gasPrice() async throws -> EthereumQuantity
-
-  func estimateGas(
-    invocation: SolidityInvocation,
-    from: EthereumAddress?,
-    gas: EthereumQuantity?,
-    value: EthereumQuantity?
-  ) async throws -> EthereumQuantity
-
-  func call(_ invocation: SolidityInvocation) async throws -> [String: Any]
-
-  func sendRawTransaction(
-    transaction: EthereumSignedTransaction
-  ) async throws -> EthereumData
-
-  func getLogs(
-    addresses: [EthereumAddress]?,
-    topics: [[EthereumData]]?,
-    fromBlock: EthereumQuantityTag,
-    toBlock: EthereumQuantityTag
-  ) async throws -> [EthereumLogObject]
-
-  func subscribeToLogs(
-    addresses: [EthereumAddress]?, topics: [[EthereumData]]?
-  ) -> AsyncThrowingStream<EthereumLogObject, Error>
-}
-
-extension Web3Async {
-  func getTransactionCount(
-    for address: EthereumAddress, block: EthereumQuantityTag = .latest
-  ) async throws -> EthereumQuantity {
-    try await getTransactionCount(for: address, block: block)
-  }
-
-  func estimateGas(
-    invocation: SolidityInvocation,
-    from: EthereumAddress? = nil,
-    gas: EthereumQuantity? = nil,
-    value: EthereumQuantity? = nil
-  ) async throws -> EthereumQuantity {
-    try await estimateGas(invocation: invocation,
-                          from: from,
-                          gas: gas,
-                          value: value)
-  }
-
-  func getLogs(
-    addresses: [EthereumAddress]? = nil,
-    topics: [[EthereumData]]? = nil,
-    fromBlock: EthereumQuantityTag = .latest,
-    toBlock: EthereumQuantityTag = .latest
-  ) async throws -> [EthereumLogObject] {
-    try await getLogs(addresses: addresses,
-                      topics: topics,
-                      fromBlock: fromBlock,
-                      toBlock: toBlock)
-  }
-
-  func subscribeToLogs(
-    addresses: [EthereumAddress]? = nil,
-    topics: [[EthereumData]]? = nil
-  ) -> AsyncThrowingStream<EthereumLogObject, Error> {
-    subscribeToLogs(addresses: addresses, topics: topics)
-  }
-}
-
 struct Web3AsyncAdapter: Web3Async {
   private let web3: Web3
